Fixed insert tests
