+++ go/libraries/doltcore/sqle/schema_util_test.go
@@ -1,16 +1,18 @@
-// Copyright 2019 Liquidata, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ * // Copyright 2020 Liquidata, Inc.
+ * //
+ * // Licensed under the Apache License, Version 2.0 (the "License");
+ * // you may not use this file except in compliance with the License.
+ * // You may obtain a copy of the License at
+ * //
+ * //     http://www.apache.org/licenses/LICENSE-2.0
+ * //
+ * // Unless required by applicable law or agreed to in writing, software
+ * // distributed under the License is distributed on an "AS IS" BASIS,
+ * // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * // See the License for the specific language governing permissions and
+ * // limitations under the License.
+ */
 
 package sqle
 
@@ -220,6 +222,7 @@ func CompressRow(sch schema.Schema, r row.Row) row.Row {
 	var itag uint64
 	compressedRow := make(row.TaggedValues)
 
+	// TODO: this is probably incorrect and will break for schemas where the tag numbering doesn't match the declared order
 	sch.GetAllCols().IterInSortedOrder(func(tag uint64, col schema.Column) (stop bool) {
 		if val, ok := r.GetColVal(tag); ok {
 			compressedRow[itag] = val
